# angular-ui/bao-loc-api.openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/v1/solutions/{solutionId}":{"get":{"tags":["Solutions"],"summary":"Retrive a specific solution. User must own this solution or have BAOLOC_ADMIN role.","operationId":"retrieveSolution","parameters":[{"name":"solutionId","in":"path","required":true,"schema":{"type":"long"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolutionResponse"}},"application/xml":{"schema":{"$ref":"#/components/schemas/SolutionResponse"}}}}}},"put":{"tags":["Solutions"],"summary":"Update a specific solution. User must own this solution or have BAOLOC_ADMIN role.","operationId":"updateSolution","parameters":[{"name":"solutionId","in":"path","required":true,"schema":{"type":"long"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolutionUpdateRequest"}},"application/xml":{"schema":{"$ref":"#/components/schemas/SolutionUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}},"application/xml":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Solutions"],"summary":"Delete a specific solution. User must own this solution or have BAOLOC_ADMIN role.","operationId":"deleteSolution","parameters":[{"name":"solutionId","in":"path","required":true,"schema":{"type":"long"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}},"application/xml":{"schema":{"type":"object"}}}}}}},"/v1/solutions":{"get":{"tags":["Solutions"],"summary":"Retrieve solutions for current user","operationId":"retrievePlayerSolutions","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SolutionResponse"}}},"application/xml":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SolutionResponse"}}}}}}},"post":{"tags":["Solutions"],"summary":"Add a solution to current user collection","operationId":"createSolution","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolutionUpdateRequest"}},"application/xml":{"schema":{"$ref":"#/components/schemas/SolutionUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}},"application/xml":{"schema":{"type":"object"}}}}}}},"/v1/users/{username}/roles":{"get":{"tags":["Users"],"operationId":"getUserRoles","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRolesDto"}}}}}}}},"components":{"schemas":{"SolutionUpdateRequest":{"required":["x1","x2","x3","x4","x5","x6","x7","x8","x9"],"type":"object","properties":{"x1":{"maximum":9,"minimum":1,"type":"integer","format":"int64"},"x2":{"maximum":9,"minimum":1,"type":"integer","format":"int64"},"x3":{"maximum":9,"minimum":1,"type":"integer","format":"int64"},"x4":{"maximum":9,"minimum":1,"type":"integer","format":"int64"},"x5":{"maximum":9,"minimum":1,"type":"integer","format":"int64"},"x6":{"maximum":9,"minimum":1,"type":"integer","format":"int64"},"x7":{"maximum":9,"minimum":1,"type":"integer","format":"int64"},"x8":{"maximum":9,"minimum":1,"type":"integer","format":"int64"},"x9":{"maximum":9,"minimum":1,"type":"integer","format":"int64"}}},"UserRolesDto":{"type":"object","properties":{"authorities":{"type":"array","items":{"type":"string"}}}},"SolutionResponse":{"required":["id","x1","x2","x3","x4","x5","x6","x7","x8","x9"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"x1":{"type":"integer","format":"int64"},"x2":{"type":"integer","format":"int64"},"x3":{"type":"integer","format":"int64"},"x4":{"type":"integer","format":"int64"},"x5":{"type":"integer","format":"int64"},"x6":{"type":"integer","format":"int64"},"x7":{"type":"integer","format":"int64"},"x8":{"type":"integer","format":"int64"},"x9":{"type":"integer","format":"int64"}}}}}}